New test suite and a bunch of fixes in hqq-database.
Namely, initialization of the hqq-database finally works as intended,
with categories resulting from items added to the hqq-database at
initialization as an array if provided (NB the fill-pointer should be
set properly), but without severely fucking up the category list in
the event of an empty array being provided.  All it took was nesting
an "if" inside that "unless," I guess! I'll be damned.